requirements.txt
asgiref==3.5.0
boto3==1.20.46
botocore==1.23.46
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.1
django-allauth==0.47.0
django-storages==1.12.3
idna==3.3
jmespath==0.10.0
mysqlclient==2.1.0
oauthlib==3.2.0
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
